refactor: 레벨2 야구게임 객체 파일 분리
